[home][Android] Exclude diagnostics screen code on Android
