+++ home/screens/DiagnosticsScreen/LocationDiagnosticsScreen.tsx
@@ -17,9 +17,9 @@ import {
 } from 'react-native';
 import MapView, { Polyline } from 'react-native-maps';
 
-import NavigationEvents from '../components/NavigationEvents';
-import Button from '../components/PrimaryButton';
-import Colors from '../constants/Colors';
+import NavigationEvents from '../../components/NavigationEvents';
+import Button from '../../components/PrimaryButton';
+import Colors from '../../constants/Colors';
 
 const STORAGE_KEY = 'expo-home-locations';
 const LOCATION_UPDATES_TASK = 'location-updates';
