+++ home/screens/DiagnosticsScreen/index.android.tsx
@@ -1,0 +1,3 @@
+export function DiagnosticsStackScreen() {
+  return null;
+}
